goodbye _sha_compare
